+++ src/main/java/com/bookreum/dev/domain/club/repository/ClubRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.bookreum.dev.domain.club.entity.ClubEntity;
 
 @Repository
-public interface  ClubRepository extends JpaRepository<ClubEntity, Long> {
+public interface  ClubRepository extends JpaRepository<ClubEntity, Integer> {
 
 }
